refactor(azure): share SQL adapter helpers
