+++ packages/api/src/adapter-azure/AzureSqlAdapter.ts
@@ -18,6 +18,16 @@ import type {
     SqlTable,
 } from '../cloud-spi/types'
 import {MssqlDataClient, type SqlDataClient, type SqlDataConnection} from './MssqlDataClient'
+import {
+    filterBySearch,
+    isValidServerName,
+    numberValue,
+    rawStringValue,
+    recordValue,
+    stringValue,
+    tagList,
+    withApiVersion,
+} from './sqlAdapterUtils'
 
 const API_VERSION = '2021-11-01'
 const RESOURCE_ID_PREFIX = 'sql-server:'
@@ -145,7 +155,7 @@ export class AzureSqlAdapter implements CloudServiceAdapter {
     }
 
     private sqlFetch(path: string, init: RequestInit, emptyOnNotFound = false): Promise<Response | null> {
-        return this.client.fetch(withApiVersion(path), {
+        return this.client.fetch(withApiVersion(path, API_VERSION), {
             ...init,
             headers: {
                 'accept': 'application/json',
@@ -194,11 +204,6 @@ function sqlServerPath(serverName: string): string {
     return `${sqlServersPath()}/${encodeURIComponent(serverName)}`
 }
 
-function withApiVersion(path: string): string {
-    const separator = path.includes('?') ? '&' : '?'
-    return `${path}${separator}api-version=${API_VERSION}`
-}
-
 function sqlServerName(id: string): string {
     const name = isSqlServerResourceId(id) ? id.slice(RESOURCE_ID_PREFIX.length) : id
     if (!name) throw new ValidationError('Azure SQL server id is required')
@@ -239,41 +244,9 @@ function toSqlServerResource(server: AzureSqlRecord): CloudResource {
     }
 }
 
-function filterBySearch(resources: CloudResource[], search?: string): CloudResource[] {
-    const normalized = search?.trim().toLowerCase()
-    if (!normalized) return resources
-    return resources.filter((resource) => resource.name.toLowerCase().includes(normalized))
-}
-
-function tagList(tags?: Record<string, string>): Array<{key: string; value: string}> {
-    return Object.entries(tags ?? {}).map(([key, value]) => ({key, value}))
-}
-
-function stringValue(value: unknown): string {
-    return typeof value === 'string' ? value.trim() : ''
-}
-
-function rawStringValue(value: unknown): string {
-    return typeof value === 'string' ? value : ''
-}
-
-function recordValue(value: unknown): Record<string, unknown> | null {
-    return value && typeof value === 'object' && !Array.isArray(value)
-        ? value as Record<string, unknown>
-        : null
-}
-
-function numberValue(value: unknown): number | null {
-    return typeof value === 'number' && Number.isFinite(value) ? value : null
-}
-
 function integerValue(value: unknown): number | null {
     if (typeof value === 'number' && Number.isSafeInteger(value)) return value
     if (typeof value !== 'string' || !/^\d+$/.test(value)) return null
     const parsed = Number(value)
     return Number.isSafeInteger(parsed) ? parsed : null
-}
-
-function isValidServerName(value: string): boolean {
-    return value.length <= 63 && /^[a-z0-9](?:[a-z0-9-]*[a-z0-9])?$/.test(value)
 }
